Prevent duplicate PROGMEM strings
Normally, the compiler automatically merges any duplicate string
constants in a program. However, when using the F() macro, this no
longer works, since F() allocates a hidden static variable, whose
address is returned, so each time F() is used a new variable is
allocated in PROGMEM and any duplicate strings occur multiple times.
This is magnified when F() is used in a template function, since a new
variable will be allocated for each template instantiation.

This commit resolves this by allocating some strings in global PROGMEM
variables, which can then be referenced when used.

When compiling the parse example on the Arduino Uno, this saves 920
bytes of flash.