Make sure that recursive ParsedData methods are inlined
The ParsedData class contains a list of fields and method to parse and
print those fields. This class is implemented using variadic templates
that inherit from each other recursively, each adding one field. When
actually parsing or printing, methods are called recursively working
their way up through the inheritance chain, parsing or printing one
field at the time.

Before, this resulted in a very deep call stack for the parse method.
Since every one of those methods allocates a ParseResult object
containing the parsed value on the stack, this would result in a lot of
stack memory usage when a lot of fields were parsed. This issue
prevented running the parse example on an Arduino Uno.

Ideally, this would be solved by the compiler by applying tail recursion
optimization. For the applyEach method, this happens since the previous
commit, but for the parse method this optimization is not applied for
unknown reasons. However, even with tail recursion, there is significant
overhead for each the pre- and postamble of the recursively called
methods, both in flash space and runtime. Since none of these methods is
every called directly, other than through this recursion, performance is
greatly increased by forcing all of these calls to be inlined.

The result is one big parse or applyEach method, that contains the code
to parse or iterate each field in turn.

When compiling the parse example for the Arduino Uno, this saves 4538
bytes of flash. Additionally, this allows to run the parse example on
the Arduino Uno unmodified.

Note that the same thing also applies to the ParsedData constructors and
destructors. However, applying the same fix there is not so easy, since
adding a specially named "inline" version of a constructor or destructor
is not possible. For the constructor, an extra constructor argument
could be used to differentiate, but destructors cannot take any
arguments. Making all constructors and destructors always_inline would
work, but again introduce significant overhead if multiple instances of
a ParsedData object are used. Since tail recursion optimization is
applied for constructors and destructors, this is left as it is.