Pass argument to ParsedData::applyEach by reference
This allows a functor that keeps state, prevents making a copy from
being made on every recursive call, and allows tail recursion to be
applied.

In a simple case, where the applied functor class has no members and no
destructor (like the Printer in the examples), the extra copies were
already optimized away, so this change will actually increase the code
size a bit (because the actual references have to be passed on). The
tail recursion improvement still applies in this case.
l