Move unit variables into struct
Global const(expr) variables have implicit internal linkage, meaning
each cpp file that includes us will have its own copy of the variable.
Since we take the address of these variables (passing it as a template
argument), this caused a compiler warning. By putting these in a struct,
we can define the value in the header file, but allocate the storage for
them in the cpp file, preventing this warning.