Fix underflow in parser for decimal parsing
In this case, the max_decimals could wrap around to UINT_MAX when the
amount of decimals after the '.' was more than 3. This would cause the
loop to keep running for ~UINT_MAX iterations, in practice hanging the
whole parser. This patch fixes this behavior and cleanly errors out if
the amount of decimals is > 3.

This will result in the following error message:

```
1-0:1.8.1(.00385.6
              ^
Missing unit
```

Example of an invalid message: `1-0:1.8.1(.00385.6`

This fixes issue #30.