Fix reinterpret_cast error on newer gcc versions
Newer gcc versions correctly enforce a C++ language limitation that
forbids using reinterpret_cast in constant expressions. This commit
works around this by replacing the `name` attribute with a `get_name()`
method that delays the cast until the time of use. Note that the method
will be inlined in practice, so there is no additional overhead.

For compatibility with existing code, the name attribute is replaced by
a NameConverter object that can be implicitly converted to `const
__FlashStringHelper *`. This should keep some code working, but not
everything (e.g. conversion to String can no longer be done implicitly,
see #43).

In the future, the name attribute should probably be completely removed,
and replaced by a `name()` method (which is a bit more elegant than the
current `get_name()`)

This fixes #36.